tests/mediacapture-streams/MediaStream-tests-manual.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test MediaStream interface with Color Stream</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://w3c.github.io/mediacapture-main/#mediastream">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<meta name="flags" content="interact">

<p class="instructions">Use a test device with color camera(embedded or external).</p>
<p class="instructions">When prompted, accept to share your color stream.</p>

<script>

//modify the test title
promise_test(t => {
  let allowedCharacters = /^[\u0021\u0023-\u0027\u002A-\u002B\u002D-\u002E\u0030-\u0039\u0041-\u005A\u005E-\u007E]*$/;
  return navigator.mediaDevices.getUserMedia({video: true})
  .then(stream => {
    assert_true(stream.id.length === 36, "the media stream id has 36 characters");
    assert_regexp_match(stream.id, allowedCharacters, "the media stream id uses the set of allowed characters");
  });
}, "Tests that a MediaStream is created by getUserMedia() has a correct id");

promise_test(t => {
  return navigator.mediaDevices.getUserMedia({video: true})
  .then(stream => {
    let track = stream.getVideoTracks()[0];
    assert_equals(track, stream.getTrackById(track.id), "getTrackById returns track of given id");
    assert_equals(stream.getTrackById(track.id + "foo"), null, "getTrackById of inexistant id returns null");
  })
  .catch(error => {
    assert_unreached("Unexpected getUserMedia error: " + error);
  });
}, "Tests that MediaStream.getTrackById works as expected");
//need move to MediaSteam audio file and modify audivideoo keyword
promise_test(t => {
  return navigator.mediaDevices.getUserMedia({audio: true})
  .then(stream => {
    assert_true(stream instanceof MediaStream, "getUserMedia success callback comes with a MediaStream object");
    assert_equals(stream.getAudioTracks().length, 1, "the media stream has exactly one audio track");
    assert_equals(stream.getAudioTracks()[0].kind, "audio", "getAudioTracks() returns a sequence of tracks whose kind is 'audio'");
    assert_equals(stream.getVideoTracks().length, 0, "the media stream has zero video track");
  });
}, "Tests that a MediaStream with exactly one audio track is returned");

promise_test(t => {
  return navigator.mediaDevices.getUserMedia({video: true})
  .then(stream => {
    assert_true(stream instanceof MediaStream, "getUserMedia success callback comes with a MediaStream object");
    assert_equals(stream.getAudioTracks().length, 0, "the media stream has zero audio track");
    assert_equals(stream.getVideoTracks().length, 1, "the media stream has exactly one video track");
    assert_equals(stream.getVideoTracks()[0].kind, "video", "getAudioTracks() returns a sequence of tracks whose kind is 'video'");
  });
}, "Tests that a MediaStream with at least one video track is returned");

promise_test(t => {
  return navigator.mediaDevices.getUserMedia({video: true, audio: true})
  .then(stream => {
      let stream1 = new MediaStream();
      assert_not_equals(stream.id, stream1.id, "Two different MediaStreams have different ids");

      let stream2 = new MediaStream(stream);
      assert_not_equals(stream.id, stream2.id, "A MediaStream constructed from another have different ids");
      let audioTrack1 = stream.getAudioTracks()[0];
      let videoTrack = stream.getVideoTracks()[0];
      assert_equals(audioTrack1, stream2.getAudioTracks()[0], "A MediaStream constructed from another share the same audio track");
      assert_equals(videoTrack, stream2.getVideoTracks()[0], "A MediaStream constructed from another share the same video track");

      let stream3 = new MediaStream([audioTrack1]);
      assert_equals(stream3.getTrackById(audioTrack1.id), audioTrack1, "a non-ended track gets added via the MediaStream constructor");
      let audioTrack2 = audioTrack1.clone();
      audioTrack2.addEventListener("ended", t.unreached_func("ended event should not be fired by MediaStreamTrack.stop()."));
      audioTrack2.stop();
      assert_equals(audioTrack2.readyState, "ended", "a stopped track is marked ended synchronously");

      let stream4 = new MediaStream([audioTrack2, videoTrack]);
      assert_equals(stream4.getTrackById(audioTrack2.id), audioTrack2, "an ended track gets added via the MediaStream constructor");
      assert_equals(stream4.getTrackById(videoTrack.id), videoTrack, "a non-ended track gets added via the MediaStream constructor even if the previous track was ended");

      let stream5 = new MediaStream([audioTrack2]);
      assert_equals(stream5.getTrackById(audioTrack2.id), audioTrack2, "an ended track gets added via the MediaStream constructor");
      assert_false(stream5.active, "a MediaStream created using the MediaStream() constructor whose arguments are lists of MediaStreamTrack objects that are all ended, the MediaStream object MUST be created with its active attribute set to false");
  });
}, "Tests that a MediaStream constructor follows the algorithm set in the spec");
//change get video stream first to get audio stream first.
promise_test(t => {
  let audio, video;
  return navigator.mediaDevices.getUserMedia({audio: true})
  .then(audio_stream => {
    audio = audio_stream;
    return navigator.mediaDevices.getUserMedia({video: true})
    .then(video_stream => {
      video = video_stream;
      video.getVideoTracks()[0].stop();
      assert_false(video.active, "video stream is inactive after stopping its only video track");
      assert_true(audio.active, "audio stream is active");
      video.addTrack(audio.getAudioTracks()[0]);
      video.removeTrack(video.getVideoTracks()[0]);
    });
  });
}, "Tests that adding a track to an inactive MediaStream is allowed");
//delete add function return void
promise_test(t => {
  let audio, video;
  return navigator.mediaDevices.getUserMedia({audio: true})
  .then(audio_stream => {
    audio = audio_stream;
    return navigator.mediaDevices.getUserMedia({video: true})
    .then(video_stream => {
      video = video_stream;
      assert_equals(video.getAudioTracks().length, 0, "video mediastream starts with no audio track");
      video.addTrack(audio.getAudioTracks()[0]);
      assert_equals(video.getAudioTracks().length, 1, "video mediastream has now one audio track");
      video.addTrack(audio.getAudioTracks()[0]);
      assert_equals(video.getAudioTracks().length, 1, "video mediastream still has one audio track"); // If track is already in stream's track set, then abort these steps.
      audio.onaddtrack = t.step_func(function () {
        assert_unreached("onaddtrack is not fired when the script directly modified the track of a mediastream");
      });
      assert_equals(audio.getVideoTracks().length, 0, "audio mediastream starts with no video track");
      audio.addTrack(video.getVideoTracks()[0]);
      assert_equals(audio.getVideoTracks().length, 1, "audio mediastream now has one video track");
    });
  });
}, "Tests that adding a track to a MediaStream works as expected");
//delete remove function return void
promise_test(t => {
  let audio, video;
  return navigator.mediaDevices.getUserMedia({audio: true})
  .then(audio_stream => {
    audio = audio_stream;
    return navigator.mediaDevices.getUserMedia({video: true})
    .then(video_stream => {
      video = video_stream;
      video.onremovetrack = function () {
        assert_unreached("onremovetrack is not triggered when removal of track is triggered by the script itself");
      };
      assert_equals(video.getVideoTracks().length, 1, "video mediastream starts with one video track");
      video.removeTrack(video.getVideoTracks()[0]);
      assert_equals(video.getVideoTracks().length, 0, "video mediastream has no video track left");
      video.removeTrack(audio.getAudioTracks()[0]); // should not throw
    });
  });
}, "Tests that a removal from a MediaStream works as expected");

</script>
